+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/crds/Dummy.java
@@ -15,33 +15,15 @@
  */
 package io.fabric8.kubernetes.examples.crds;
 
+import io.fabric8.kubernetes.api.model.KubernetesResource;
 import io.fabric8.kubernetes.api.model.Namespaced;
-import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.client.CustomResource;
+import io.fabric8.kubernetes.model.annotation.Group;
+import io.fabric8.kubernetes.model.annotation.Version;
 
-/**
- */
-public class Dummy extends CustomResource implements Namespaced {
-  private DummySpec spec;
-
-
-  @Override
-  public String toString() {
-    return "Dummy{" +
-        "apiVersion='" + getApiVersion() + '\'' +
-        ", metadata=" + getMetadata() +
-        ", spec=" + spec +
-        '}';
-  }
-
-  public DummySpec getSpec() {
-    return spec;
-  }
-
-  public void setSpec(DummySpec spec) {
-    this.spec = spec;
-  }
-
-  @Override
-  public ObjectMeta getMetadata() { return super.getMetadata(); }
+@Version(Dummy.VERSION)
+@Group(Dummy.GROUP)
+public class Dummy extends CustomResource<DummySpec, KubernetesResource> implements Namespaced {
+  public static final String GROUP = "demo.fabric8.io";
+  public static final String VERSION = "v1";
 }